requirements.txt
Django
djangorestframework
python-dotenv
psycopg2

